// src/stories/Input/styles.module.scss.d.ts
declare const styles:
  & Readonly<{ "input": string }>
  & Readonly<{ "input": string }>
  & Readonly<{ "input": string }>
  & Readonly<{ "isInvalid": string }>
  & Readonly<{ "wrapper": string }>
  & Readonly<{ "label": string }>
  & Readonly<{ "error": string }>
  & Readonly<{ "characterCount": string }>
  & Readonly<{ "labelAndCountContainer": string }>
  & Readonly<{ "requiredApostrophe": string }>
;
export default styles;
//# sourceMappingURL=./styles.module.scss.d.ts.map